configuración de entorno de node para correr el frontend
