+++ .gitignore
@@ -1,0 +1,61 @@
+# Node.js
+node_modules/
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+# Logs
+logs/
+*.log
+*.log.*
+
+# Environment variables
+.env
+
+# Operating system files
+.DS_Store
+Thumbs.db
+
+# Editor-specific files
+.idea/
+.vscode/
+*.swp
+
+# Build files
+/dist
+/build
+/out
+
+# Temporary files
+*.tmp
+*.temp
+*.pid
+
+# Public uploads or temporary content (if applicable)
+public/uploads/
+public/tmp/
+
+# Specific IDE config files
+*.iml
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
+
+# Coverage directory for testing
+coverage/
+
+# Miscellaneous
+*.bak
+*.orig
+*.log
+
+# Optional npm cache directory
+.npm/
+
+# Optional eslint cache
+.eslintcache
+
+# Optional prettier cache
+.prettiercache
