+++ package.json
@@ -1,0 +1,27 @@
+{
+  "name": "login-html",
+  "version": "1.0.0",
+  "description": "Esta es una simple aplicación de login construida con HTML, CSS y JavaScript. Permite al usuario ingresar sus credenciales, validarlas, y redirigirlo o mostrar mensajes de éxito o error según el caso. La aplicación también utiliza el almacenamiento local (localStorage) para guardar la información del usuario al momento de un inicio de sesión exitoso.",
+  "main": "app.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon server.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/StevenAngelSofka/Login-html.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/StevenAngelSofka/Login-html/issues"
+  },
+  "homepage": "https://github.com/StevenAngelSofka/Login-html#readme",
+  "dependencies": {
+    "express": "^4.21.2"
+  },
+  "devDependencies": {
+    "nodemon": "^3.1.9"
+  }
+}
